requirements.txt
neptune
pytorch-lightning
torchvision
scikit-learn
segmentation-models-pytorch
matplotlib
opencv-python
natsort
python-dotenv